--- t-rdfGraphMerger/src/main/java/cz/cuni/mff/xrg/uv/transformer/graphmerge/GraphMergeConfig_V1.java
@@ -1,12 +1,0 @@
-package cz.cuni.mff.xrg.uv.transformer.graphmerge;
-
-/**
- * DPU's configuration class.
- */
-public class GraphMergeConfig_V1 {
-
-    public GraphMergeConfig_V1() {
-
-    }
-
-}
