+++ t-rdfGraphMerger/src/main/java/eu/unifiedviews/plugins/transformer/rdfgraphmerger/RdfGraphMergerConfig_V1.java
@@ -1,0 +1,12 @@
+package eu.unifiedviews.plugins.transformer.rdfgraphmerger;
+
+/**
+ * DPU's configuration class.
+ */
+public class RdfGraphMergerConfig_V1 {
+
+    public RdfGraphMergerConfig_V1() {
+
+    }
+
+}
